Make suggestions module injectable through component API
